Split currency symbols from essential patterns
Separate symbol mappings (placeholders, pattern configurations, and default pattern selection) into a dedicated CurrencySymbolsV1 marker and struct (symbols.rs), leaving CurrencyEssentialsV1 purely for core layout rules (patterns and indices).

- Create CurrencySymbolsV1 with pattern_config_map, placeholders, and default_pattern_config in components/experimental/src/dimension/provider/currency/symbols.rs.
- Extract CurrencySymbolsV1 in provider/source/src/currency/symbols.rs from CLDR (symbol and symbol-narrow).
- Update CurrencyEssentialsV1 in essentials.rs to only retain patterns and indices.
- Update CurrencyFormatterData::Essential and formatters in formatter.rs & compact_formatter.rs to require and use both CurrencyEssentialsV1 and CurrencySymbolsV1.
- Update provider registry (provider/registry/src/lib.rs) and regenerate test and baked data for CurrencySymbolsV1 and CurrencyEssentialsV1.

TAG=agy
CONV=d044eeab-1ec7-4391-9cf0-a6d3d9e6edf0
