+++ provider/data/experimental/stubdata/mod.rs
@@ -60,6 +60,7 @@ include!("units_names_duration_outlier_v1.rs.data");
 include!("narrow_quarter_relative_v1.rs.data");
 include!("units_names_volume_core_v1.rs.data");
 include!("script_display_names_v1.rs.data");
+include!("currency_symbols_v1.rs.data");
 include!("narrow_minute_relative_v1.rs.data");
 include!("digital_duration_data_v1.rs.data");
 include!("locale_names_region_short_v1.rs.data");
@@ -162,6 +163,7 @@ macro_rules! impl_data_provider {
         impl_narrow_quarter_relative_v1!($provider);
         impl_units_names_volume_core_v1!($provider);
         impl_script_display_names_v1!($provider);
+        impl_currency_symbols_v1!($provider);
         impl_narrow_minute_relative_v1!($provider);
         impl_digital_duration_data_v1!($provider);
         impl_locale_names_region_short_v1!($provider);
