+++ provider/source/src/currency/mod.rs
@@ -8,3 +8,4 @@ pub(crate) mod essentials;
 pub(crate) mod extended;
 pub(crate) mod fractions;
 pub(crate) mod patterns;
+pub(crate) mod symbols;
